requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-docx==1.1.2
pydantic==2.9.2
